requirements.txt
Django==4.2.29
djangorestframework==3.15.2
django-cors-headers==4.7.0
